+++ bin/ekko_cli.dart
@@ -1,7 +1,7 @@
-import 'package:get_cli/get_cli.dart';
+import 'package:ekko_cli/ekko_cli.dart';
 
 Future<void> main(List<String> arguments) async {
-  final command = GetCli(arguments).findCommand();
+  final command = EkkoCli(arguments).findCommand();
   if (command != null && command.validate()) {
     await command.execute();
   }