+++ lib/commands.dart
@@ -1,5 +1,5 @@
-import 'package:get_cli/commands/help/help.dart';
-import 'package:get_cli/commands/version/version.dart';
+import 'package:ekko_cli/commands/help/help.dart';
+import 'package:ekko_cli/commands/version/version.dart';
 
 import 'commands/create/screen/screen.dart';