+++ lib/commands/create/create.dart
@@ -1,10 +1,10 @@
-import 'package:get_cli/get_cli.dart';
+import 'package:ekko_cli/ekko_cli.dart';
 
 mixin CreateMixin {
-  final _args = GetCli.arguments;
+  final _args = EkkoCli.arguments;
 
   bool get withOn => _args.length > 3 && _args[3] == 'on';
-  String get on => GetCli.arguments[4];
+  String get on => EkkoCli.arguments[4];
 
   String get name {
     return _args[1].split(':')[1];