+++ lib/commands/help/help.dart
@@ -1,6 +1,6 @@
-import 'package:get_cli/commands.dart';
-import 'package:get_cli/common/utils/logger/logger.dart';
-import 'package:get_cli/core/command.dart';
+import 'package:ekko_cli/commands.dart';
+import 'package:ekko_cli/common/utils/logger/logger.dart';
+import 'package:ekko_cli/core/command.dart';
 
 class HelpCommand extends Command {
   @override