+++ lib/core/sample.dart
@@ -1,6 +1,6 @@
 import 'dart:io';
-import 'package:get_cli/commands/create/screen/functions/replace_as_expected.dart';
-import 'package:get_cli/common/utils/logger/logger.dart';
+import 'package:ekko_cli/commands/create/screen/functions/replace_as_expected.dart';
+import 'package:ekko_cli/common/utils/logger/logger.dart';
 import 'package:path/path.dart';
 
 abstract class Sample {