+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-## 1.0.0
+## 0.1.0
+
+- Create screen Command
+- Use: ekko create screen:<ScreeName>
+
+## 0.0.1
 
 - Initial version, created by Stagehand
+- Thanks for Stagehand to make this boilerplate, I'm using to create a CLI for my architeture.